Add plan validation for categories, credit cards, and advanced dashboard filters

+++ xpenbox-back/src/main/java/org/xpenbox/payment/dto/PlanFeatureResponseDTO.java
@@ -8,6 +8,6 @@ package org.xpenbox.payment.dto;
  */
 public record PlanFeatureResponseDTO(
     String featureCode,
-    Integer limitValue,
+    Long limitValue,
     Boolean isEnabled
 ) { }
